+++ lib/data/models/meal_aliment_model.dart
@@ -11,8 +11,8 @@ class MealAliment {
   final meal = IsarLink<Meal>();
   final aliment = IsarLink<Aliment>();
   String? dose;
-  short? quantity;
-  short? calories;
+  int? quantity;
+  int? calories;
   double? proteins;
   double? carbohydrates;
   double? sugars;
